+++ floaty_chatheads/example/lib/examples/accessibility_example.dart
@@ -155,6 +155,7 @@ class _AccessibilityExampleState extends State<AccessibilityExample> {
   @override
   void dispose() {
     _sub?.cancel();
+    FloatyChatheads.closeChatHead();
     FloatyChatheads.dispose();
     super.dispose();
   }