+++ floaty_chatheads/example/lib/examples/dashboard_example.dart
@@ -217,6 +217,7 @@ class _DashboardExampleState extends State<DashboardExample> {
     _sub?.cancel();
     _titleController.dispose();
     _bodyController.dispose();
+    FloatyChatheads.closeChatHead();
     FloatyChatheads.dispose();
     super.dispose();
   }