+++ floaty_chatheads/example/lib/examples/messenger_example.dart
@@ -224,6 +224,7 @@ class _MessengerExampleState extends State<MessengerExample> {
     _sub?.cancel();
     _controller.dispose();
     _scrollController.dispose();
+    FloatyChatheads.closeChatHead();
     FloatyChatheads.dispose();
     super.dispose();
   }