+++ floaty_chatheads/example/lib/examples/messenger_fullscreen_example.dart
@@ -189,6 +189,7 @@ class _MessengerFullscreenExampleState
   void dispose() {
     _sub?.cancel();
     _controller.dispose();
+    FloatyChatheads.closeChatHead();
     FloatyChatheads.dispose();
     super.dispose();
   }