+++ floaty_chatheads/example/lib/examples/mini_player_example.dart
@@ -193,6 +193,7 @@ class _MiniPlayerExampleState extends State<MiniPlayerExample> {
   @override
   void dispose() {
     _sub?.cancel();
+    FloatyChatheads.closeChatHead();
     FloatyChatheads.dispose();
     super.dispose();
   }