+++ floaty_chatheads/example/lib/examples/multi_chathead_example.dart
@@ -211,6 +211,7 @@ class _MultiChatheadExampleState extends State<MultiChatheadExample> {
   @override
   void dispose() {
     _dataSub?.cancel();
+    FloatyChatheads.closeChatHead();
     FloatyChatheads.dispose();
     super.dispose();
   }