+++ floaty_chatheads/example/lib/examples/notification_counter_example.dart
@@ -127,6 +127,7 @@ class _NotificationCounterExampleState
   @override
   void dispose() {
     _sub?.cancel();
+    FloatyChatheads.closeChatHead();
     FloatyChatheads.dispose();
     super.dispose();
   }